# api-gateway-user/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.Category:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  genproto.Course:
    properties:
      category:
        $ref: '#/definitions/genproto.Category'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  genproto.CourseCreateReq:
    properties:
      category_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  genproto.CourseListsRes:
    properties:
      courses:
        items:
          $ref: '#/definitions/genproto.Course'
        type: array
      pagination:
        $ref: '#/definitions/genproto.Pagination'
      total_count:
        type: integer
    type: object
  genproto.CourseUpt:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  genproto.EvaluationCreate:
    properties:
      Feedback:
        type: string
      Score:
        type: integer
      UserCourseId:
        type: string
    type: object
  genproto.EvaluationGet:
    properties:
      CreatedAt:
        type: string
      Feedback:
        type: string
      Id:
        type: string
      Score:
        type: integer
      UserCourseId:
        $ref: '#/definitions/genproto.UserCourse'
    type: object
  genproto.EvaluationList:
    properties:
      Get:
        items:
          $ref: '#/definitions/genproto.EvaluationGet'
        type: array
      Pagination:
        $ref: '#/definitions/genproto.Pagination'
      TotalCount:
        type: integer
    type: object
...
  genproto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  genproto.UserCourse:
    properties:
      course:
        $ref: '#/definitions/genproto.Course'
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/genproto.UserGetRes'
    type: object
...
  genproto.UserGetRes:
    properties:
      address:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_confirmed:
        type: boolean
      last_name:
        type: string
      occupation:
        type: string
      phone_number:
        type: string
      photo_url:
        type: string
    type: object
...
  genproto.Void:
    type: object
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /course/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Course'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get course details by ID
      tags:
      - courses
  /course/create:
    post:
      consumes:
      - application/json
      description: Create a new course with the provided details
      parameters:
      - description: Course details
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/genproto.CourseCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Void'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new course
      tags:
      - courses
  /course/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Void'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a course by ID
      tags:
      - courses
  /course/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of courses with optional filters
      parameters:
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.CourseListsRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all courses
      tags:
      - courses
  /course/update/{id}:
    put:
      consumes:
      - application/json
      description: Update the details of a specific course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Course details
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/genproto.CourseUpt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Void'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update course details by ID
      tags:
      - courses
  /evaluations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an evaluation by its ID
      parameters:
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Evaluation deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Evaluation not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Evaluation
      tags:
      - Evaluation
    get:
      consumes:
      - application/json
      description: Retrieve an evaluation by its ID
      parameters:
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Evaluation details
          schema:
            $ref: '#/definitions/genproto.EvaluationGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Evaluation not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Evaluation
      tags:
      - Evaluation
  /evaluations/create:
    post:
      consumes:
      - application/json
      description: Create a new evaluation
      parameters:
      - description: Evaluation data
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/genproto.EvaluationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Evaluation created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Evaluation
      tags:
      - Evaluation
  /evaluations/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of evaluations
      parameters:
      - description: Course ID
        in: query
        name: course_id
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Score From
        in: query
        name: score_from
        type: integer
      - description: Score To
        in: query
        name: score_to
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of evaluations
          schema:
            $ref: '#/definitions/genproto.EvaluationList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Evaluations
      tags:
      - Evaluation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"